requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
mysql-connector-python==8.3.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
openai==1.6.1
python-jose==3.3.0
requests==2.31.0
Pillow==10.1.0
flask-cors==4.0.0 